chore: formattedDate 뷰에서 분리
