+++ GimiFeedback/GimiFeedback/Presentation/FeedbackChannel/FeedbackChannelView.swift
@@ -104,12 +104,3 @@ struct FeedbackChannelView: View {
         
     }
 }
-
-extension Date {
-    fileprivate var formattedDate: String {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "ko_KR")
-        formatter.dateFormat = "yyyy년 M월 d일"
-        return formatter.string(from: self)
-    }
-}
