+++ GimiFeedback/GimiFeedback/Util/Extension/Date+Extension.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+extension Date {
+    var formattedDate: String {
+        let formatter = DateFormatter()
+        formatter.locale = Locale(identifier: "ko_KR")
+        formatter.dateFormat = "yyyy년 M월 d일"
+        return formatter.string(from: self)
+    }
+}
+
